Fix setlocale to accept also a different encoding and most of wide ch…

+++ library/include/time.h
@@ -75,7 +75,7 @@ __BEGIN_DECLS
 typedef _CLOCKID_T_ clockid_t;
 
 typedef unsigned long clock_t;
-typedef unsigned long time_t;
+typedef long long time_t;
 
 struct tm
 {
